requirements.txt
# Python 3.8.10

argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
autopep8==1.6.0
backports.zoneinfo==0.2.1
cffi==1.15.0
Django==4.0.5
django-environ==0.9.0
djangorestframework==3.13.1
mysqlclient==2.1.1
Pillow==9.1.1
pycodestyle==2.8.0
pycparser==2.21
pytz==2022.1
sqlparse==0.4.2
toml==0.10.2
